fix partial bugs and first round refactor
This problem is harder than rest binary search problem. the reason is
that the result is not inside the final [start, end] array. Also, it is
hard to derive from the final [start, end] array. Understand this
additional factor is the key to solve this problem. Refer to the last
two test cases