refactor code
When we build adjacency list graph, we don't use two loops to check
every pair of string to see if they are 1 character apart. Instead, we
make changes to current string to obtain all the strings we can reach
from current node, and see if it is in the wordList. Thus, there are
currentString.length() * 25 case we need to check for every node. This
is faster when the wordList set is large, since the check-every-pair
method need wordList.size() * currentString.length() for each node.
Otherwise, your may exceed the running time limit.

For the strings we visited, we remove it from the wordList. This way we
don't need to mark visited using another HashSet or something.

Actually, we don't even need to build the adjacency list graph
explicitly using a HashMap<String, ArrayList<String>>, since we keep all
the nodes we can reach in the queue of each level of BFS. This can be
seen as the keys of the HashMap are the strings that in the queue, and
values are the strings that satisfy the 1 character apart in the
wordList. Thus, we avoid the time cost of build map for those nodes we
don't need to visit.